utils: split into submodules and fix imports
- move remaining methods from `streamlink.utils` to `streamlink.utils.*`
  - streamlink.utils.cache.LRUCache
  - streamlink.utils.data.search_dict
  - streamlink.utils.module.load_module
  - streamlink.utils.rtmp.{escape_librtmp,rtmpparse}
  - streamlink.utils.swf.swfdecompress
- fix imports in streamlink, streamlink_cli and tests
- move utils tests into new `tests.utils` package
  - split up `test_utils` into individual test modules
  - move and rename `test_utils_*`
  - move and rename `test_localization`
  - fix up some tests
