+++ src/streamlink/plugins/youtube.py
@@ -9,7 +9,7 @@ from streamlink.plugin.api import useragents, validate
 from streamlink.plugin.api.utils import itertags
 from streamlink.stream import HLSStream, HTTPStream
 from streamlink.stream.ffmpegmux import MuxedStream
-from streamlink.utils import search_dict
+from streamlink.utils.data import search_dict
 from streamlink.utils.parse import parse_json
 
 log = logging.getLogger(__name__)
