+++ src/streamlink/utils/module.py
@@ -1,0 +1,15 @@
+from importlib.machinery import FileFinder, SOURCE_SUFFIXES, SourceFileLoader
+from importlib.util import module_from_spec
+
+
+_loader_details = [(SourceFileLoader, SOURCE_SUFFIXES)]
+
+
+def load_module(name, path=None):
+    finder = FileFinder(path, *_loader_details)
+    spec = finder.find_spec(name)
+    if not spec or not spec.loader:
+        raise ImportError(f"no module named {name}")
+    mod = module_from_spec(spec)
+    spec.loader.exec_module(mod)
+    return mod
