+++ src/streamlink/utils/swf.py
@@ -1,0 +1,8 @@
+import zlib
+
+
+def swfdecompress(data):
+    if data[:3] == b"CWS":
+        data = b"F" + data[1:8] + zlib.decompress(data[8:])
+
+    return data
