+++ tests/test_plugins.py
@@ -4,7 +4,7 @@ import unittest
 
 import streamlink.plugins
 from streamlink.plugin.plugin import Matcher, Plugin
-from streamlink.utils import load_module
+from streamlink.utils.module import load_module
 
 
 class PluginTestMeta(type):
